pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-mml-server"
version = "0.1.0"
description = "MML MCP Server - MML（Music Macro Language）を処理するModel Context Protocol (MCP)サーバー"
authors = [
    {name = "MCP MML Server Team"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
]
dependencies = [
    "mcp[cli]",
    "python-dotenv",
    "mido",
    "python-rtmidi",
    "music21",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
mcp-mml-server = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]